// src/app/shared/generated/model/productPageResult.ts
/**
 * onecx-product-store-bff
 * Backend-For-Frontend (BFF) service for onecx product store. This API provides endpoints to manage applications (technical microfrontend(s)) and product(s) as logical abstraction. A Product is a versioned cover for a collection of applications (versioned) to be used within workspaces. Microfrontends (applications) which have reference to exposed/registered modules
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ProductAbstract } from './productAbstract';


export interface ProductPageResult { 
    /**
     * The total elements in the resource (counted items in resultset).
     */
    totalElements?: number;
    /**
     * Number of page which has been queried (e.g. 2 of 10)
     */
    number?: number;
    /**
     * correlated to the input (pagesize)), how many products should be shown on a single page.
     */
    size?: number;
    /**
     * representing the total amount of pages based on the size provided in the request.
     */
    totalPages?: number;
    /**
     * result set of the given search criteria.
     */
    stream?: Array<ProductAbstract>;
}